website: Explain how to authenticate the provider when managing TFE with TFE
TFE automatically writes a .terraformrc file with a credentials block when it
does Terraform runs, so theoretically the provider could authenticate
automatically. However, the automatically provided token is a temporary "run"
token with extremely limited privileges, and in practice it cannot manage any of
this provider's resources.

But .terraformrc is still the best method when running on the CLI.