Update workspace ID to match TFE and TF
In TFE a workspace is usually presented as `organization/workspace`.

And if you want to `force-unlock` a remote state (a workspace) in TF, you have to pass `organization/workspace` as the lock ID.

So it makes sense to use the same format as the internal ID for a workspace in the `tfe_workspace` resource. Espacially when we are adding imports, as then people are required to pass that ID to the import command.