Handle interpolated cases
Something I didn't think of earlier was the fact that values for
'access' or 'permissions' could be interpolated and not known at plan
time. In a nutshell, this allows interpolation for 'access' but
restricts the ability for 'permissions' to be interpolated, as it
creates a scenario that is impossible to model with the current SDK.

As a side effect, this also means all 'permissions' arguments are
required if a 'permissions' block is provided. That's ok.