variable: use ForceNew for key changes on sensitive variables (#175)
The API does not support updating the value of sensitive variables. If someone changes the value of such a variable and attempts to apply the change the API will reject it and the apply operation will fail.

Instead of attempting to update the value we will now delete the existing variable and create a new one.