SDK2 upgrade: Remove MaxItems validation from sometimes-computed fields
This complex API behavior (https://github.com/hashicorp/terraform-provider-tfe/blob/1836332f965b90962c2b852b86baa7b9482bbcac/website/docs/r/team_access.html.markdown#L41-L52)
was bound to have _something_ to change in V2 of the SDK.

Per https://github.com/hashicorp/terraform-plugin-sdk/commit/83c83c4a1d88cce479df5a8832a31f14667a6f1d :

These schema validations are now not allowed for computed attributes, as
it is assumed that computed attributes are *always* computed; this API,
however, relies on the fact that a user can specify either "access" or
"permissions" and whichever is not specified is computed, and
vice-versa.

With this removal, it's sadly possible to specify multiple "permissions"
blocks without an error (but there should be); this will result in a
plan attempting to add a second set of permissions that results in a
no-op (as only the first item in the permissions list is used).

It's probably possible to munge further into the CustomizeDiff logic to
just return an error ourselves at plan time should there be data at
"permissions.1.*", but I'm not going to touch that right now.