SDK2 upgrade: Relax regex on Notification Configurations
The output in V2 of the SDK must have changed slightly, as the
whitespacing looks ever so different and the messages aren't anchored to
the beginning of the line as expected here, so the tests fail.

The regexs here don't need to be this strict. Search for the substring
and that's a-ok.