SDK2 upgrade: Use Terraform 0.12 for acceptance tests
In V2 of the SDK, acceptance tests are done using a real Terraform
binary and not via the inner workings of Terraform itself.

By default, the latest version of Terraform is installed and used
(unless specified via options to use a path or version)

In upgrading to the V2 SDK, it seems that we actually have some test
configurations that are not compatible with 0.13 (the current version,
as of this writing), and cause tests to fail:

=== FAIL: tfe TestAccTFEOrganizationMembershipDataSource_basic (1.60s)
    data_source_organization_membership_****.go:16: Step 1/1 error: Error running pre-apply plan:
        Error: Error retrieving organization memberships: resource not found

=== FAIL: tfe TestAccTFEWorkspaceIDsDataSource_basic (1.56s)
    data_source_workspace_ids_****.go:16: Step 1/1 error: Error running pre-apply plan:
        Error: Error retrieving workspaces: resource not found

This issue is due to a known issue with 0.13, issue number 26074 on
hashicorp/terraform, and I confirmed that even using SDKv1 these tests
fail if you use Terraform 0.13. Because of this, let's just force usage
of 0.12 (which is how things are run now) and fix these configurations
separately.