Fix remote state consumer regression for TFE
In v0.25.0, a regression was introduced when adding the new remote state
consumers feature. Unlike other relationships in the provider - which are
serialized in the same original resource document - remote state
consumers on a workspace only exist at a separate endpoint which must be
read. When the endpoint can't be read in the case of a preexisting
Terraform Enterprise installation without the feature, it causes a
failure when doing anything with the tfe_workspace resource.

An ideal scenario would be for a version check to determine whether a
call should be made - but the current tooling here is minimal. Instead,
we apply the following heuristic to make it backwards compatible for
TFE:

If a call to the remote state consumers endpoint fails and is a 404 Not
Found, assume that the only plausible reason for it is an old TFE
installation: ignore it entirely and set the computed
global_remote_state attribute to an implicit 'true' value (which
indicates the old behavior which is not configurable until you upgrade
your Terraform Enterprise installation). This is reasonable because
having reached that point, the primary call to read the workspace
succeeded, meaning you are authorized to the workspace and the
likelihood of the remote state consumers call failing from a Not Found
error for any other reason is exceedingly small.