Improve discovery and loading of credentials from Terraform config files
This provider is somewhat special, because the API token it uses is also used
by Terraform CLI itself. We can give a nicer experience on the local CLI by
auto-discovering Terraform's copy of the token (if one isn't provided in the
provider config or the `TFE_TOKEN` environment variable).

However, we weren't quite matching Terraform's credential-finding behavior:

- We were respecting the `TERRAFORM_CONFIG` environment variable for the main
  config file location, but not `TF_CLI_CONFIG_FILE`. (Terraform will use
  either, but prefers the latter since it matches the other environment variable
  names it uses.)
- We weren't ever consulting the credentials file
  (`~/.terraform.d/credentials.tfrc.json`, on unix), which is written to by
  `terraform login` and/or credential helper scripts. The HCL library
  understands the credentials JSON, so you could work around this by pointing
  `TERRAFORM_CONFIG` at the credentials file, but if there was anything else
  important in the config file you also had to set `TF_CLI_CONFIG_FILE` so that
  Terraform (but not the provider) could still find it.

This commit fixes both of those: we'll accept a non-default config file location
via either environment variable (preferring `TF_CLI_CONFIG_FILE`), and we'll
_also_ consult the credentials file if present.

If both sources include a token for the desired hostname, the provider will
follow Terraform's behavior and prefer the token from the CLI config file.