Match internal muxed provider config to the main one
The 0.26.0 release introduced features that required muxing the existing
provider built with terraform-plugin-sdk and another one built with a
lower level provider package, terraform-plugin-go. What this means is
that there are two providers which must interpret the same
configuration/settings, in one.

The SDK schema usually provides a default hostname value with
DefaultFunc, and unfortunately this behavior was not translated to an
equivalent in new lower level provider now muxed in, causing a
regression where unless a hostname is explicitly set in configuration or
TFE_HOSTNAME, the provider configuration will error.

Defaults and fallbacks which are usually handled by DefaultFunc in the
main schema are now instead handled when fetching a TFC/E client in
getClient(), so that both provider schemas fall back on the same logic
for these meta values.