Added Tfe Terraform Version resource
This commit implements the Terraform version resource block "tfe_terraform_version" which
allows practitioners to create and manage Terraform versions on TFE.

This commit also outlines two tests for this resource. It generates a mock terraform version using a random version number and SHA. The URL attribute points to hashicorp.com since the TFC API doesn't download or verify that the zip exists when the resource is created. The basic test tests against the required parameters and the full test includes the optional parameters.