Add Run Tasks Helper
This commit adds a helper functions `fetchOrganizationRunTask` and
`fetchWorkspaceRunTask` which will be used in later commits to be able
to find a Run Task by name in an organization or workspace.

This commit also adds unit tests for helper and uses uses the mocking
provided by the go-tfe module.

Later commits will migrate the rest of the project to gomocks.