Change default for `file_triggers_enabled` to false (#510)
Change default for file_triggers_enabled to false.

Why?
Current default is wrong and we're having a clause on the TFC side to account for this behavior.
New trigger types, e.g. tag based triggering, will depend on this field being set to false. We'll introduce the conflict between the new fields and file_triggers_enabled.