Add searching to tfe_oauth_client data source
To use the tfe_oauth_client data source, you must provide the OAuth
Client ID. But finding this ID is difficult, since it is not directly
exposed in the TFE/TFC user interface.

This commit adds the ability to search for an OAuth Client by its name
and/or VCS provider in combination with the organization name.

Since the 'Name' property may be nil and may not be unique per OAuth
Client, this data source will return an error if no matching OAuth
Client is found or if multiple OAuth Clients are found which match the
given parameters.