Skip update cost estimation acctest.
This particular test needs to support a kind of workflow that is
currently not supported by the SDK test harness. Toggling cost
estimation is a paid feature which requires an organization that is not
on the free plan. Remove the test skip once either the SDK supports
importing a resource before applying a configuration, or remove the test
entirely once the `cost_estimation` attribute is removed from this resource.