Fetch outputs steps is failing from time to time
We are not completely sure what is causing it, though there is some speculations.
We are removing the name field in case is conflicting with the job id called tests-summarize.
We are also adding some additional context to some errors in the go program for fetch outputs to help debug in the future.
Finally, adding some additional options to the workflow bash that runs with only -e option by default. I will remove this after we see the step no longer failing with the unauthorized error.