Remove `sidebar_current:` from all docs yaml frontmatter
This frontmatter key does nothing, and any effort spent adding it to new files
or keeping it up to date is tragic. Let's yank it so newer contributors aren't
tempted to mess with it!

Background story: This used to be a hack to make the sidebar navs work properly
on the old Middleman-based Terraform.io website. I eventually replaced that hack
with a better hack, later we stopped hosting provider docs on terraform.io at
all, and then some years later we moved that site off Middleman entirely anyway.