REWRITE tfe_variable resource with terraform-plugin-framework
This commit adds a complete re-implementation of tfe_variable. It keeps most
existing behavior, but fixes a notable bug (#839) where we were unable to
respect `ignore_changes` for sensitive variable values. This was impossible to
address in the v2 SDK, but is incredibly straightforward with the new framework.

Although the effort required for this rewrite wouldn't necessarily be justified
for a single bug fix, it:

- Helps validate our hypothesis that the framework can indeed address the
recurring scourge of "bugs not fixable due to inability to distinguish zero from
absent".

- Gets the framework muxed in and ready to rock and roll, so we can immediately
start implementing new resource types in it.

- Establishes some initial patterns, practices, and examples for using the new
framework within this particular provider.

Anyway, here's some interesting sidenotes about this rewrite:

- The migration docs for the new framework were invaluable:
https://developer.hashicorp.com/terraform/plugin/framework/migrating

- Likewise the working code for the "HashiCups" teaching provider:
https://github.com/hashicorp/terraform-provider-hashicups-pf

- Note that I omitted some logic in most of the existing CRUD methods that
fetched the workspace prior to trying to manage the variable. This logic dated
from the time when the workspace_id argument used the `org/ws_name` format
instead of an external ID; it's no longer necessary now that we have a usable ID
right off the bat.

- Attributes that have a `Default` must also be `Computed`, apparently. I guess
I can see it.

- Most of the go-tfe struct types (update options, etc.) are only constructed by
one CRUD method, but most the methods need to convert a go-tfe value to a model
that we can use to update the state; thus, that's the logic I pulled out into
helper funcs. Conveniently, that's also the point where we need to take some
extra care to make sure we're preserving our last-known information about the
value of a sensitive variable, so we're able to keep that logic centralized too.