HEY ACTUALLY, don't bother fetching the workspace before going for it with the var!
Turns out, that logic I was following from the old impl dated back to when 1.
go-tfe's variables functions had very different API and took a workspace struct
embedded in them, and 2. variable resources had an organization/ws_name pair!
So... let's reap the benefit of all that work we did to migrate to external IDs.