Delete the old implementation of tfe_variable
Note: This also blows away the old variableV0 schema migration tests, and I'm
not actually sure yet how to test the new style of state migrator. The only idea
I've had so far is to do something similar to the regression acceptance test
included in this PR, and use ExternalProviders to pull in the last version of
the provider before the `workspace_id` changed in the `tfe_variable` schema.
HOWEVER, said schema change happened in late 2019, and I'm not even sure a
provider that old would work properly with a modern TFC?

Perhaps we keep this around as a best-effort, but don't bother testing it
because of ever-diminishing likelihood anyone will invoke it. At any rate, at
least it provides an example of how to implement a state upgrader in the new
setup.