Add acceptance test for state stability in tfe_variable rewrite
This should verify that the new resource implementation is a safe drop-in
replacement for the old one, causing no state changes when run against existing
resource state.

The utility of this test is expected to diminish over time, and we'll probably
need to delete it the next time we make a notable change to the workspace or
variable set resources. All things pass!