Docs: Explain use cases, flag race condition, add waits to examples
- The examples are implicitly demonstrating wait behavior on the parent apply,
and it basically always makes sense to await on a destroy, so add the requisite
waits.

- There's currently a race condition that requires force_delete for destroys.
Mention that in a call-out, and add force_delete to the examples so they aren't
broken out of the box.

- The use cases for this resource are kind of complex, so add some guidance to
help people get on its level (so they don't try to use it for managing random
operational lifecycle runs 😝).