Fix flaky team and organization token tests
Each team or org may only have one token at a time. Managing two at once is
illegal and is expected to cause an error.

These tests appear to have been copy-pasted from tests that were explicitly
testing the provider's handling of already extant tokens, but were not intended
to exercise those features and were not written in a way that guaranteed the
order of token creation. Thus, they are destined to flake.

This commit removes the illegal extra tokens from the configs, which appears to
match the behavior the tests were intended to exercise.