Fix Terraform 1.6 compat by finishing GetMetadata implementation
Providers implemented using the `terraform-plugin-go` library above version 0.19
or so must implement GetMetadata(), as part of supporting an optional memory
usage improvement. See
https://github.com/hashicorp/terraform-plugin-sdk/issues/1234 for the start of a
breadcrumb trail with more details about this whole effort.

Almost nobody actually implements providers using terraform-plugin-go, but we
do! It's one of our THREE muxed-together providers, because we once needed to do
something that was impossible in the SDK before the new framework existed. It's
used for ONE data source, `tfe_outputs`.

This function was added (with a blank return value) in
https://github.com/hashicorp/terraform-provider-tfe/pull/1046 and no one noticed
a problem, presumably because acceptance tests were running with a Terraform
version that didn't exercise the new memory-saving code path, instead just
calling GetProviderSchema however many times.

But, Terraform 1.6 *does* exercise it, and it turns out a blank value causes it
to bail out early with a not-found for the tfe_outputs data source.

This commit adds our lonely data source to the low-level provider's metadata,
which fixes the issue. Longer-term, we should... probably migrate that data
source to the framework and delete that whole third provider.