Fix flaky `TestAccTFEVariableSetsDataSource_full`
This test was checking that the varset was connected to the project, but that
connection is managed by a separate resource that was not a declared or inferred
dependency of the data source. So, random chance determined whether the varset
had been connected to the project yet by the time the data source was read.

The fix is to declare the dependency.