Fix nil pointer dereference segfault in tfe_variable resource
If go-tfe returns a non-nil err, you can't ever treat the `*tfe.Variable`
pointer it returns as though it's valid. But in the Update path for this
resource, I forgot to do an early return on client error, and thus we'd
immediately try to dereference that variable pointer and do a segfault.

Easiest way to reproduce this bug is to apply a config that creates a
non-sensitive variable, manually create a second variable, then try to change
the `key` of the config variable to conflict with the key of the manual
variable.

All the other crud methods were fine — Create and Read both had appropriate
early returns, and Delete doesn't need it because it doesn't do anything after
appending to diags. I'm guessing I must have implemented Update after Delete and
gotten confused.