r/tfe_workspace: Fix panic when `trigger_prefixes = [""]`
A customer reported this panic via support. Sebastian clued me in that this is
an extremely cursed known behavior in sdkv2, where it munges a list of strings
containing one empty string into a list containing a single nil, so you must do
a type assert on the inner elements.

Note that the Update function already did exactly this (around line 655 or so)!
So there was no panic if you changed trigger_prefixes to the fatal value after
the resource already existed.