Refactor testing helpers
Previously some of the test only helpers did not have the _test.go name which
implies it's only used in the acceptance tests. This commit renames those files.

This commit also extracts the subscription update process and updated it as it
fails in the local development. Specifically, there may be 2 or more tokens
used to administer a TFC instance which have different admin scopes. This commit
now encapsulates that logic inside a SubscriptionUpdater "service" class which
can be used to build up the request to update a subscription.

This service will be used in later commits where we need to test different
tiers an organization is in. This is typically only a TFC concern, not TFE.