Refactor Run Task tests for HMAC Key
Previously the acceptance tests assumed that the Run Task service test fixture
would allow any old HMAC Key. However this is not the case in the CI
environment. Instead we now allow the acceptance test suite to specifiy the
HMAC Key, and use manual tests for testing how to change this write-only
sensitive value.