Rename tag_bindings to tags, add ignore_additional_tags
Rather than use the nomenclature defined by the API, we instead simplify
the attribute field to simply be called tags. This commit also adds
ignore_additional_tags which acts as a safeguard against tags added
outside of Terraform configuration. Since tags are updated in-place,
this ensures that the user does not inadvertently remove tags
their configuration is unaware of.