tests: replace all remaining sdk provider factories with framework muxed provider factories (#1669)
* tests: add ctor for framework provider with a default org

* tests: add default org muxed factory and save global test client reference

* tests: first pass migrating to framework muxed provider factory

This commit is intended to only have one change, spread across many
files.

* tests: introduce muxed provider with default org

* tests(resource_tfe_workspace_test): update testAccCheckTFEWorkspaceExists

This is the first commit focused on resource_tfe_workspace_test.
It removes the dependency on the SDK provider from the
testAccCheckTFEWorkspaceExists check.

* tests(resource_tfe_workspace_test): update testAccCheckTFEWorkspaceDestroy

This is the second commit focused on resource_tfe_workspace_test.

It removes the dependency on the SDK provider from the
testAccCheckTFEWorkspaceDestroy check.

* tests(resource_tfe_workspace_test): migrate remaining tests to use muxed provider factory

* tests(resource_tfe_no_code_module_test): remove remaining dependencies on sdk provider

* tests: replace remaining uses of sdk provider with global client instance

This commit is intended to have the same change spread across many
files.

* tests(resource_tfe_workspace_settings_test): remove sdk provider reference from testAccCheckTFEWorkspaceSettingsDestroy check

This commit is focused on resource_tfe_workspace_settings_test.

* tests: clean up remaining sdk test provider refs

* tests: ensure global test client is set for all muxed providers

* tests: fix org default logic for muxedProvidersWithDefaultOrganization