[IPL-7245] Fix OAuth Token ID updates not registering (#1631)
* [IPL-7245] Ensure `oauth_token_id` updates register when performing a `terraform apply`

This resolves the issue where when updating the `oauth_token_id`, `terraform apply` would have to be performed twice before to confirm the change in the state file. The first apply would perform the change on the server but would not register the change in the state file as expected.

* tests: use muxed provider factory

* chore: update changelog

---------

Co-authored-by: Chris Trombley <chris.trombley@hashicorp.com>