Use token ID as resource ID when multiple tokens
Previously, when a team only had a single token, it was sufficient to have
the ID of the token be set to the team ID. Now that we support multiple
team tokens, we should use the token ID instead. We will continue to
use the team ID for descriptionless tokens so that it is backwards
compatible, though.