fix: prevent duplicate set element error in project data source
The GET /workspaces API call is ordered by run.created-at.
It's possible that between retrieving the first page, and the next page
that a workspace will appear on both pages.
This then ultimately causes a 'Duplicate Set Element' error
in the TF plan.

This fix converts storing the API response as 2 lists to a single
map, with the key being the workspace ID ensuring we overwrite any
duplicates.